Now ignoring nodes with "connect = no".
This fixes MON-11928.

The plugin will now check for the "flags" value of each node and filter out any
node where the value indicates that the merlin configuration has "connect = no".

There is also a new option: "-c", "--include-connect-no", which will ignore such
flags and test all nodes not on the ignore list.